to be fixed string as a func result, parameter

+++ docs/C++May2016/HW4/infoArraysAPI.h
@@ -2,7 +2,7 @@
 #include <fstream>
 
 namespace infoArrays {
-typedef unsigned int ID;
+typedef long long ID;
 
 bool isFileExist(const char* fileName)
 {	bool res;
